Moved MakeResourceRequest over to a method on RequestLine
