+++ playground/writable_route.go
@@ -22,7 +22,7 @@ func (route *ReadWriteRoute) Route(requested *http.RequestLine) http.Request {
 		return nil
 	}
 
-	return http.MakeResourceRequest(requested, route.Resource)
+	return requested.MakeResourceRequest(route.Resource)
 }
 
 type ReadWriteResource interface {
